requirements.txt
requests
beautifulsoup4
natsort
python-dateutil

flask
flask-sqlalchemy
flask-migrate
Flask-Pydantic

gunicorn
sentry-sdk[flask]
python-dotenv
pydantic

pytest
pytest-flask
pytest-mock

pre-commit